+++ src/models/user.ts
@@ -1,15 +1,20 @@
 import { ObjectId } from "mongodb";
 import { userRole, userSex } from "../types/user";
-export default class User {
-    constructor(
-        public id: number,
-        public createdAt: Date,
-        public name: string,
-        public email: string,
-        public pwd: string,
-        public sex: userSex,
-        public role: userRole,
-        public _id?: ObjectId,
-        public blocked?: number[]
-    ) {}
+
+export interface BlockedUser {
+    id: number;
+    date: Date;
+    reason: string;
+}
+
+export default interface User {
+    _id?: ObjectId;
+    id: number;
+    createdAt: Date;
+    name: string;
+    email: string;
+    pwd: string;
+    sex: userSex;
+    role: userRole;
+    blocked?: BlockedUser[];
 }