+++ src/routes/me/blocked.ts
@@ -1,7 +1,7 @@
 import { FastifyInstance, FastifyPluginOptions } from "fastify";
 import { usersCl } from "../../common";
 import verifyUser from "../../lib/auth/verify";
-import User from "../../models/user";
+import User, { BlockedUser } from "../../models/user";
 
 export default function (
     fastify: FastifyInstance,
@@ -14,19 +14,19 @@ export default function (
 
         const blocked = ((
             await usersCl.findOne({ id: user.id }, { projection: { _id: 0, blocked: 1 } })
-        )?.blocked || []) as number[];
+        )?.blocked || []) as BlockedUser[];
 
-        console.log(blocked);
-
-        const blocklist = (await usersCl
-            .find({
-                id: { $in: blocked },
-            })
-            .project({ _id: 0, id: 1, name: 1, sex: 1, role: 1 })
+        const usersBlocked = (await usersCl
+            .find(
+                { id: { $in: blocked.map((b) => b.id) } },
+                { projection: { _id: 0, id: 1, name: 1, sex: 1, role: 1 } }
+            )
             .toArray()) as User[];
 
         res.send(
-            blocked.map((id) => blocklist.find((i) => i.id === id)).filter((i) => i)
+            blocked
+                .map((b) => ({ ...b, ...usersBlocked.find((u) => u.id === b.id) }))
+                .filter((i) => i.name && i.id)
         );
     });
     done();