+++ src/routes/thread/index.ts
@@ -6,6 +6,8 @@ import create from "./create";
 import comment from "./comment";
 import category from "./category";
 import hidden from "../../plugins/hidden";
+import star from "./actions/star";
+import unstar from "./actions/unstar";
 
 export default (
     fastify: FastifyInstance,
@@ -18,6 +20,8 @@ export default (
     fastify.register(images);
     fastify.register(create);
     fastify.register(category);
+    fastify.register(star);
+    fastify.register(unstar);
     fastify.register(comment, { prefix: "/:id/comment" });
     done();
 };